(ADD) error handling

+++ app/services/BuyProductService.py
@@ -1,4 +1,5 @@
 from utils.tools import selDB, insDB
+from utils import tools
 from base.baseModels import BuyProduct
 
 class BuyProducts:
@@ -9,13 +10,16 @@ class BuyProducts:
     def get():
         query = "SELECT * FROM buy_products;"
         products = selDB(query)
-        return products
+        return tools.payloadSuccess(products, 200)
   
     def getById(buy_id):
         query = "SELECT * FROM buy_products where id=%s"
         values = (buy_id, )
-        products = selDB(query, values)
-        return products
+        product = selDB(query, values)
+        if product:
+            return tools.payloadSuccess(product, 200)
+        else:
+            return tools.payloadError("Product not found", 404)
     
     def post(buyProduct: BuyProduct):
         query = "insert into buy_products (product_id, price, rate_product, shop, quantity, order_id) values (%s, %s, %s, %s, %s, %s)"
@@ -23,12 +27,16 @@ class BuyProducts:
         aux = insDB(query, values)
         query = "update products set quantity = quantity + "+str(buyProduct.quantity)+" where id="+str(buyProduct.product_id)+""
         aux = insDB(query, values=None)
-        return aux
+        return tools.payloadSuccess(buyProduct, 201)
   
     def put(product): pass
 
     def delete(buy_id):
-        query = "delete from buy_products where id=%s"
-        values = (buy_id,)
-        insDB(query, values)
-        return True
+        if not "error" in BuyProducts.getById(buy_id):
+            query = "delete from buy_products where id=%s"
+            values = (buy_id,)
+            insDB(query, values)
+            return {"message": "delete with successful"}
+        else:
+            return tools.payloadError("Buy not found", 404)
+        return
